adds pretty metrics graphs
I was pondering how to present the data collected and decided
graphite+grafana in docker was worth a shot.  Turns out it was dead
simple to set up. This adds a docker-compose that starts up those two
plus an erlang container to run the code in. After starting it,
connect to the erlang container with

    docker exec -it -w /app erl-snmp-profiler_snmp_profiler_1 bash

You can build and run the app there using the shell scripts. Will that
default container name change? It might, so watch out.

There are some basic metrics being emitted as examples. Not sure if we
want to keep them, but after running, visit http://localhost:3000 to
see a default dashboard with one panel showing the metrics just
emitted.