adds to metrics
This starts to think about the right metrics to output. It turns out
the counter needs to actually keep count itself. Grafana doesn't seem
to have a way to aggregate individual events of "count 1" like I was
trying to do. I think it should work fine to use ETS to keep track of
the counter value. It just means they all reset to 0 each time you
start the program again.