changes --dc to --datafile
I've got things working for a single switch. Now I'm thinking about
how to do multiple, and I think it's better to refer to a list of
switches as a datafile than a dc. It can be for a dc or not. I think
this will keep things a little simpler.