changes snmp listen port
Port 5000 was in use on a server I ran this on. Changing to a more
random port will hopefully avoid that in the future. Ideally, I'd grab
an unused ephemeral port at startup and use that, but it would entail
a good bit of work to get the port in all the right places.