adds licensing
There doesn't seem to be any reason this couldn't be made a public
repo. It would make it simpler to distribute to users, since you could
write a script to download and extract the archive for them without
authentication needed. Therefore, this adds a license and minimal
contributing guide to make it more open-source-friendly.